Updated some ints to doubles to handle decimals

+++ formulas.hpp
@@ -1,15 +1,15 @@
 #include <iostream>
 
 
-int addition(double a, double b){
+double addition(double a, double b){
     return a + b;
 }
-int subtraction(double a, double b){
+double subtraction(double a, double b){
     return a-b;
 }
-int product(double a, double b){
+double product(double a, double b){
     return a*b;
 }
-int division(double a, double b){
+double division(double a, double b){
     return a/b;
 }
